Experiment to remove the WorldPanel to render the video and instead write directly to the projection texture
